requirements.txt
openai
python-dotenv
python-box
dirigera
spotipy


SpeechRecognition
numpy
pandas
tabulate
duckdb